+++ CHANGELOG.md
@@ -14,7 +14,8 @@ below this description) is likely unreleased.
 
 ## Changed
 
-- Renamed `command.ExistingFlags` to `commamd.FlagMap`. The plural name is awkward when adding a one-element `FlagMap` is perfectly ok (and nice when you don't want to keep typing the name...)
+- Changed `warg.VersionCommand` to `warg.VersionCommandMap` and `warg.ColorFlag` to `warg.ColorFlagMap`. This somewhat standardizes the name.
+- Renamed `command.ExistingFlags` to `commamd.FlagMap` and `warg.ExistingGlobalFlags` to `warg.GlobalFlagMap`. The plural name is awkward when adding a one-element `FlagMap` is perfectly ok (and nice when you don't want to keep typing the name...)
 - `version` is now a parameter to `warg.New` instead of the (now removed)
 `warg.OverrideVersion`. Changed because I ALWAYS want to set the version and I
 think thats common for real-world CLIs. If passed an empty string, warg will